update Makefile
- add stamp file
- add pattern target

+++ pre-build.sh
@@ -4,5 +4,10 @@ set -e
 set -u
 set -o pipefail
 
+if [[ ! -f /etc/debian_version ]]; then
+  echo "minideb can currently only be built on debian based distros, aborting..."
+  exit 1
+fi
+
 apt-get update
 apt-get install -y debootstrap debian-archive-keyring jq dpkg-dev
